pyproject.toml
[tool.poetry]
name = "quantumfinanceai"
version = "0.1.0"
description = "Quantum Finance AI"
authors = ["Your Name <your@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.1.3"
djangorestframework = "3.15.2"
django-cors-headers = "4.4.0"
django-filter = "24.3"
djangorestframework-simplejwt = "5.4.0"
psycopg = { version = "^3.1", extras = ["binary"] }
python-dotenv = "1.0.0"
python-decouple = "3.8"
python-dateutil = "2.9.0.post0"
groq = "0.3.0"
openai = "1.30.5"
Pillow = "10.4.0"
gunicorn = "22.0.0"
whitenoise = "6.6.0"
setuptools = "70.0.0"
dj-database-url = "2.1.0"
matplotlib = "3.8.0"
reportlab = "4.0.4"
celery = "5.3.4"
redis = "5.0.1"

